fixed adding to cart

+++ SAE_S5/sqldata/tables.sql
@@ -70,7 +70,7 @@ create table BILLETS
     reference         varchar,
     dateHoraireSoiree timestamptz,
     typeTarif         varchar,
-    prix              int,
+    prix              float,
     foreign key (ID_acheteur) references USERS (ID),
     foreign key (reference) references SOIREES (ID)
 );
@@ -110,7 +110,7 @@ CREATE TABLE PANIERS
     "iduser"    uuid                   NOT NULL,
     "nbplaces"  integer                NOT NULL,
     "categorie" character varying(255) NOT NULL,
-    "montant"   integer                NOT NULL,
+    "montant"   float              NOT NULL,
     foreign key (idsoiree) references soirees (id),
     foreign key (iduser) references users (id),
     primary key (idsoiree, iduser, categorie)
